ACTIVITY SELECTION PROBLEM
The Activity Selection Problem is an optimization problem where the goal is to select the maximum number of non-overlapping activities from a set, each with a start and finish time. The challenge is to choose activities such that no two selected activities overlap in time.

Problem Definition:
Given a set of n activities, each with a start time (start[i]) and finish time (finish[i]), the task is to select the maximum number of activities that do not overlap. An activity can be selected if its start time is greater than or equal to the finish time of the previously selected activity.

Solution Approach:
The Greedy Algorithm is used to solve this problem:

1. Sort the activities based on finish time.
2. Select the first activity and then select any subsequent activity that starts after or when the previous one finishes.
3. Repeat the process until all activities are considered.

Applications:
This problem applies to areas like:

Job scheduling
Resource allocation
Event planning

Conclusion:
The Activity Selection Problem is efficiently solved using a greedy approach, ensuring the maximum number of non-overlapping activities are selected.

Applications:
This problem applies to areas like:
1. Job scheduling
2. Resource allocation
3. Event planning

Conclusion:
The Activity Selection Problem is efficiently solved using a greedy approach, ensuring the maximum number of non-overlapping activities are selected.